% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_medRxiv_data.R
\name{analyze_medrxivr}
\alias{analyze_medrxivr}
\title{Analyze medRxiv data}
\usage{
analyze_medrxivr(from_date, to_date)
}
\arguments{
\item{from_date}{Start date in "YYYY-MM-DD" format.}

\item{to_date}{End date in "YYYY-MM-DD" format.}
}
\value{
A ggplot object displaying the number of preprints in each category.
}
\description{
This function imports data from the medRxiv API for a specific time period
and creates a bar plot showing the number of preprints in each category.
}
\examples{
\dontrun{
analyze_medrxivr("2024-01-01", "2024-01-31")
}

}
