% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medrxivr.analysis-package.R
\docType{package}
\name{medrxivr.analysis-package}
\alias{medrxivr.analysis}
\alias{medrxivr.analysis-package}
\title{medrxivr.analysis: What the Package Does (Title Case)}
\description{
designed to facilitate the analysis of preprint data from the medRxiv API. It provides functions to import data from the medRxiv API, process it, and create visualizations to explore trends in preprint publications.
}
\keyword{internal}
